refactor(report): colocate tests with their modules
Follow-up to the per-renderer split: each submodule now owns its tests
in a #[cfg(test)] mod tests block, the standard Rust convention.

Shared `sample()` fixture moved to a small src/report/test_support.rs.
Module-local helpers (delta_entry, entry, render_delta_pr_to_string,
render_delta_pr_with_links) live next to the tests that use them.

src/report/tests.rs is gone; the test re-exports added to the
dispatcher in the previous commit are dropped (no longer needed —
each test mod sees its own module via super::*).
