+++ src/report/per_crate.rs
@@ -105,3 +105,99 @@ pub(crate) fn write_per_crate_markdown(
     writeln!(out)?;
     Ok(())
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::path::PathBuf;
+
+    fn entry(
+        crate_name: Option<&str>,
+        function: &str,
+        crap: f64,
+    ) -> CrapEntry {
+        CrapEntry {
+            file: PathBuf::from("src/lib.rs"),
+            function: function.into(),
+            line: 1,
+            cyclomatic: 1.0,
+            coverage: Some(100.0),
+            crap,
+            crate_name: crate_name.map(|s| s.to_string()),
+        }
+    }
+
+    #[test]
+    fn crate_rollups_aggregate_per_crate() {
+        let entries = vec![
+            entry(Some("alpha"), "a1", 1.0),
+            entry(Some("alpha"), "a2", 35.0), // crappy at threshold 30
+            entry(Some("beta"), "b1", 5.0),
+        ];
+        let rollups = crate_rollups(&entries, 30.0);
+        assert_eq!(rollups.len(), 2);
+        assert_eq!(rollups[0].name, "alpha");
+        assert_eq!(rollups[0].total, 2);
+        assert_eq!(rollups[0].crappy, 1);
+        assert_eq!(rollups[1].name, "beta");
+        assert_eq!(rollups[1].total, 1);
+        assert_eq!(rollups[1].crappy, 0);
+    }
+
+    #[test]
+    fn crate_rollups_ignore_untagged_entries() {
+        // Kills: dropping the `if let Some(name)` guard would produce a phantom
+        // empty-name row; keeping the guard correctly skips untagged entries.
+        let entries = vec![
+            entry(None, "untagged", 5.0),
+            entry(Some("alpha"), "a1", 1.0),
+        ];
+        let rollups = crate_rollups(&entries, 30.0);
+        assert_eq!(rollups.len(), 1);
+        assert_eq!(rollups[0].name, "alpha");
+    }
+
+    #[test]
+    fn crate_rollups_crappy_uses_strict_above() {
+        // Kills: replacing `>` with `>=` in the crappy count.
+        let entries = vec![
+            entry(Some("alpha"), "exactly", 30.0),
+            entry(Some("alpha"), "above", 30.1),
+        ];
+        let rollups = crate_rollups(&entries, 30.0);
+        assert_eq!(
+            rollups[0].crappy, 1,
+            "exactly-at-threshold must NOT count as crappy"
+        );
+    }
+
+    #[test]
+    fn has_crate_data_detects_any_tagged_entry() {
+        let untagged = vec![entry(None, "x", 1.0), entry(None, "y", 2.0)];
+        let one_tagged = vec![entry(None, "x", 1.0), entry(Some("alpha"), "y", 2.0)];
+        assert!(!has_crate_data(&untagged));
+        assert!(has_crate_data(&one_tagged));
+    }
+
+    #[test]
+    fn write_per_crate_human_noop_when_no_crate_data() {
+        let entries = vec![entry(None, "x", 1.0)];
+        let mut buf = Vec::new();
+        write_per_crate_human(&entries, 30.0, &mut buf).unwrap();
+        assert!(
+            buf.is_empty(),
+            "no per-crate output when no entry has crate_name"
+        );
+    }
+
+    #[test]
+    fn write_per_crate_markdown_emits_gfm_table() {
+        let entries = vec![entry(Some("alpha"), "a1", 1.0)];
+        let mut buf = Vec::new();
+        write_per_crate_markdown(&entries, 30.0, &mut buf).unwrap();
+        let s = String::from_utf8(buf).unwrap();
+        assert!(s.contains("## Per-crate summary"));
+        assert!(s.contains("| Crate | Functions | Crappy |"));
+        assert!(s.contains("| alpha | 1 | 0 |"));
+    }
+}
